refactor : 예외 클래스 분리하기
