+++ src/test/java/roomescape/MissionStepTest.java
@@ -11,7 +11,6 @@ import java.util.Map;
 
 import static org.hamcrest.Matchers.is;
 
-
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class MissionStepTest {
